add profiler

+++ util/SimpleDataLoader.py
@@ -18,4 +18,4 @@ class SimpleDataSet(Dataset):
             mean = data_range / 2
             std = 1
             data = np.random.normal(mean, scale=std, size=data_set_size)
-            self.data = data.astype(np.float)
+            self.data = data.astype(np.int64)
